refactor: split guide topics into modules
