+++ packages/stim-cli/src/guide/index.ts
@@ -1,0 +1,26 @@
+import agent from './agent.ts';
+import facts from './facts.ts';
+import metro from './metro.ts';
+import logs from './logs.ts';
+import errors from './errors.ts';
+import lifecycle from './lifecycle.ts';
+import cleanup from './cleanup.ts';
+import settings from './settings.ts';
+
+interface GuideTopic {
+  summary: string;
+  body: () => string;
+}
+
+const TOPICS: Record<string, GuideTopic> = {
+  agent,
+  facts,
+  metro,
+  logs,
+  errors,
+  lifecycle,
+  cleanup,
+  settings,
+};
+
+export default TOPICS;
